backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlmodel==0.0.22
psycopg[binary,pool]==3.2.1
redis==5.0.7
celery==5.4.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
